Create UniqueCondition class
