+++ src/InsertEdit.php
@@ -213,7 +213,7 @@ class InsertEdit
 
         $meta = $this->dbi->getFieldsMeta($result[$keyId]);
 
-        [$uniqueCondition] = Util::getUniqueCondition($meta, $rows[$keyId], true);
+        $uniqueCondition = (new UniqueCondition($meta, $rows[$keyId], true))->getWhereClause();
 
         return (bool) $uniqueCondition;
     }
@@ -864,7 +864,7 @@ class InsertEdit
         $meta = $this->dbi->getFieldsMeta($res);
         // must find a unique condition based on unique key,
         // not a combination of all fields
-        [$uniqueCondition] = Util::getUniqueCondition($meta, $row, true);
+        $uniqueCondition = (new UniqueCondition($meta, $row, true))->getWhereClause();
         if (! $uniqueCondition) {
             return;
         }
